Bit of refactoring + Comments
