+++ packages/rainbowkit-wallets/src/lib/inputsWallet.tsx
@@ -3,6 +3,9 @@ import { createRoot } from 'react-dom/client'
 import { CreateWalletModal } from '../components/CreateWalletModal'
 import localStorageWallet from './localStorageWallet'
 
+/**
+ * This class is used to create a wallet from a form
+ */
 export class inputsWallet extends localStorageWallet {
   private data: any
   private cancel: boolean = false
